+++ db/01_schema/02_domains/electrictransmission_combined_nominal_voltage_ll.sql
@@ -1,6 +1,6 @@
 -- TABLE: electrictransmission_combined_nominal_voltage_ll
 CREATE TABLE IF NOT EXISTS domains.electrictransmission_combined_nominal_voltage_ll (
-    code smallint PRIMARY KEY,
+    code bigint PRIMARY KEY,
     description varchar(64),
     details varchar(2000)
 );
