+++ db/01_schema/02_domains/electrictransmission_high_voltage_nominal_voltage.sql
@@ -1,6 +1,6 @@
 -- TABLE: electrictransmission_high_voltage_nominal_voltage
 CREATE TABLE IF NOT EXISTS domains.electrictransmission_high_voltage_nominal_voltage (
-    code smallint PRIMARY KEY,
+    code bigint PRIMARY KEY,
     description varchar(64),
     details varchar(2000)
 );
